refactor: coupon 패키지 분리
