+++ src/main/java/cart/dao/CouponRepository.java
@@ -1,6 +1,6 @@
 package cart.dao;
 
-import cart.domain.Coupon;
+import cart.domain.coupon.Coupon;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
