+++ src/main/java/cart/dto/OrderResponse.java
@@ -1,6 +1,7 @@
 package cart.dto;
 
 import cart.domain.Order;
+import cart.dto.coupon.CouponResponse;
 
 import java.time.LocalDateTime;
 import java.util.List;
